build: add clang tidy to enforce modern c++ usage (#591)
# Pull Request

## Description

To ensure that the codebase follows good principles like modern usage of
cpp, I added clang tidy configuration file with an automatic check
inside CI.

## Related Issues (Put "None" if there are no related issues)

close #553 

## Type of Change
Please delete options that are not relevant.

- Build/CI configuration change

## Changes Made
List the main changes in this PR:
- Added clang tidy config file
- Add a gate in CI to check the code conformity to clang tidy

## Testing
Describe the tests you ran to verify your changes. Please delete options
that are not relevant.
- Unit tests pass (`xmake test`)

### Test Environment
- OS: macOS
- Compiler: Clang

## Screenshots/Videos (Put "None" if there are no related issues)

None

## Documentation
Please delete options that are not relevant.

- No documentation changes are required

## Checklist (Don't delete any options)
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published

## Breaking Changes (Put "None" if there are no related issues)

None

## Additional Notes (Put "None" if there are no related issues)

None

+++ src/engine/src/scheduler/RelativeTimeUpdate.cpp
@@ -18,8 +18,8 @@ void Engine::Scheduler::RelativeTimeUpdate::RunSystems()
 {
     _bufferedTime += this->_core.GetResource<Engine::Resource::Time>()._elapsedTime;
     auto ticks = static_cast<unsigned int>(_bufferedTime / _tickRate);
-    float remainder = _bufferedTime - ticks * _tickRate;
-    _bufferedTime -= ticks * _tickRate;
+    float remainder = _bufferedTime - static_cast<float>(ticks) * _tickRate;
+    _bufferedTime -= static_cast<float>(ticks) * _tickRate;
 
     for (unsigned int i = 0; i < ticks; i++)
     {
